requirements.txt
# Nifty Options Intelligence Platform - Requirements

# Core Data Analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# Configuration
pyyaml>=6.0

# Market Data APIs
yfinance>=0.2.28

# Visualization
plotly>=5.18.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Dashboard
streamlit>=1.30.0

# Optional: Machine Learning (for advanced predictions)
scikit-learn>=1.3.0

# Optional: PDF Reports
# reportlab>=4.0.0

# Optional: Messaging
# python-telegram-bot>=20.0
